requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==4.2.2
django-cleanup==9.0.0
django-jalali-date==1.1.3
django-widget-tweaks==1.5.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
pillow==10.4.0
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
psycopg2-binary==2.9.10
environs==14.1.1